Corrected ssim and psnr cast

+++ cnn-deblur/models/ms_deblur_wgan.py
@@ -205,10 +205,10 @@ class MSDeblurWGAN(WGAN):
 
         # Compute metrics
         ssim_metric = tf.image.ssim(sharp_pyramid[0],
-                                    predicted_pyramid[0],
+                                    tf.cast(predicted_pyramid[0], dtype='bfloat16'),
                                     max_val=2.)
         psnr_metric = tf.image.psnr(sharp_pyramid[0],
-                                    predicted_pyramid[0],
+                                    tf.cast(predicted_pyramid[0], dtype='bfloat16'),
                                     max_val=2.)
         real_l1_metric = tf.abs(tf.ones_like(real_logits) - real_logits)
         fake_l1_metric = tf.abs(-tf.ones_like(fake_logits) - fake_logits)
@@ -248,10 +248,10 @@ class MSDeblurWGAN(WGAN):
 
         # Compute metrics
         ssim_metric = tf.image.ssim(sharp_pyramid[0],
-                                    predicted_pyramid[0],
+                                    tf.cast(predicted_pyramid[0], dtype='bfloat16'),
                                     max_val=2.)
         psnr_metric = tf.image.psnr(sharp_pyramid[0],
-                                    predicted_pyramid[0],
+                                    tf.cast(predicted_pyramid[0], dtype='bfloat16'),
                                     max_val=2.)
         real_l1_metric = tf.abs(tf.ones_like(real_logits) - real_logits)
         fake_l1_metric = tf.abs(-tf.ones_like(fake_logits) - fake_logits)
